+++ src/source.files/AgAppAdditionalUsers.ts
@@ -1,0 +1,12 @@
+import { UserRoles } from "./defaults";
+
+export interface AdditionalUserFormat {
+  userName: string;
+  actualName: string;
+  email: string;
+  permissionsData: UserRoles;
+}
+
+export interface AdditionalUsersReturn {
+  additionalUsers: AdditionalUserFormat[];
+}